refactor: organize source files inside `src/test` directory
