+++ src/index.ts
@@ -1,3 +1,3 @@
-export { testSolutions } from "./test/index.js";
 export { RawTestSchema, readRawTestSchema } from "./schema/raw.js";
 export { searchSolutions } from "./search.js";
+export { testSolutions } from "./test.js";
